fix(web): split agent meta into client-safe module to keep node deps off the browser bundle
