+++ shared/package.json
@@ -17,6 +17,7 @@
     "./agents": "./src/agents/base.ts",
     "./agents/claude-code": "./src/agents/claude-code.ts",
     "./agents/registry": "./src/agents/registry.ts",
+    "./agents/meta": "./src/agents/meta.ts",
     "./blocklist": "./src/blocklist.ts",
     "./draft-send": "./src/draft-send.ts",
     "./crypto": "./src/crypto.ts",
